Created a Canvas and Toolbar component
