+++ src/App.scss
@@ -9,22 +9,5 @@
     gap: 1rem;
 }
 
-.wrapper {
-    position: relative;
-}
-
-#draw-canvas {
-    background-color: transparent;
-    position: absolute;
-    width: 90vw;
-    height: 80vh;
-}
-
-#display-canvas {
-    background-color: #FFF;
-    width: 90vw;
-    height: 80vh;
-}
-
