+++ src/Components/GlobalButtons.scss
@@ -35,4 +35,21 @@
     margin-inline-end: 3.75rem;
 }
 
+.wrapper {
+    position: relative;
+}
+
+#draw-canvas {
+    background-color: transparent;
+    position: absolute;
+    width: 90vw;
+    height: 80vh;
+}
+
+#display-canvas {
+    background-color: #FFF;
+    width: 90vw;
+    height: 80vh;
+}
+
